src/backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-yasg==1.21.15
mysqlclient==2.2.8
PyMySQL==1.1.2
python-dotenv==1.2.2